Preserve insertion order in hashtables.
This should mean that iteration over objects will result in it returning
keys and values in the order they were inserted. This will also preserve the
ordering of object keys loaded.

The modifications turn the hashtable into something similar to the Java
class java.util.LinkedHashMap; it remains a hashtable, but an additional
linked list will maintain order of objects added.

Note that setting an existing key does *not* modify the insertion order;
to get that behavior, you must delete the key then set it.