Introduce JSON_LITERAL
JSON_LITERAL is a variant of JSON_STEAL. When used,
the char * is not copied and no deallocation is performed.

usage:

value = json_string_nocheck_ex("2.0",sizeof("2.0")-1,JSON_LITERAL);

maybe defining a macro would make it easer to use.

fix a bug in json_string_set_nocheck_ex() where len was not
reinitialized.

Improve test_equal so that it would have caught the fixed bug.

Signed-off-by: Olivier Langlois <olivier@olivierlanglois.net>