Renamed config.h -> jansson_private_config.h for CMake
This is because it's really easy to get a name collission if compiling
Jansson as a subproject in a larger CMake project. If one project includes
several subprojects each having their own config.h, this will cause the
wrong file to be loaded. Not sure if it's trivial to change this for
autoconf as well.