add auto_json_decref macro
gcc and clang implement the compiler attribute cleanup() [1] which lets
you add a function call that is invoked as soon as a variable goes out
of scope.

This feature improves code where you retrieve a json_t object at one
point and you're not interested to keep it over a longer period of time.

[1] https://gcc.gnu.org/onlinedocs/gcc/Common-Variable-Attributes.html#index-g_t_0040code_007bcleanup_007d-variable-attribute-3485