Avoid -Wformat due to unsupported %lld printf conversion on WIN32
If jansson is built on Windows (specifically, with MinGW 7.3.0), it produces
the following warnings:

    C:/Users/ci_user/jansson/examples/simple_parse.c:93:30: warning: unknown
        conversion type character 'l' in format [-Wformat=]

    C:/Users/ci_user/jansson/examples/simple_parse.c:106:29: warning: unknown
        conversion type character 'l' in format [-Wformat=]

Plus there are more additional warnings caused by code being treated
incorrectly because of these two.

It means it can't recognize "%lld" format specifier, which is, for Windows,
should be "%I64d". It has already been done for JSON_INTEGER_FORMAT, and I did
the same way for new macro, JSON_LONG_LONG_FORMAT.

So this commit fixes this behavior.

Further info:
https://sourceforge.net/p/mingw/bugs/2190/